fix debug rider not stop when process acad running

+++ CADPythonShell/postbuild.ps1
@@ -5,17 +5,27 @@ Write-Host "Project Directory:" $ProjectDir
 Write-Host "Configuration Name Current:" $Configuration
 $bundle = "C:\ProgramData\Autodesk\ApplicationPlugins\CADPythonShell.bundle\"
 $content = "PackageContents.xml"
+$AutodeskProcessName = "acad"
 if($Configuration -match "Debug"){
-	Write-Host "************Start Create Folder And Check File ＼（＾.＾）／"
-	if(Test-Path $bundle){
-		Write-Host "Exits Path, So Remove All File Exits"
-		Remove-Item ($bundle) -Recurse
-		Write-Host "Removed All File Exist"
+	$proc = Get-Process $AutodeskProcessName -ErrorAction SilentlyContinue
+	if($proc)
+	{
+		Write-Host "Warning : Please Close Autocad Or Civil3D To Auto Copy Resouce And Debug ＼（＾_＾）／"
 	}
-	Write-Host "************ Start Copy New File"
-	xcopy ($SolutionDir + $content) $bundle /Y
-	xcopy ($ProjectDir + "*.*") $bundle /Y /I /E /R
-	Write-Host "************ Oh my got ! Copy Complete! Chuongmep.com ＼（＾_＾）／"
+	else
+	{
+		Write-Host "************Start Create Folder And Check File ＼（＾.＾）／"
+		if(Test-Path $bundle){
+			Write-Host "Exits Path, So Remove All File Exits"
+			Remove-Item ($bundle) -Recurse
+			Write-Host "Removed All File Exist"
+		}
+		Write-Host "************ Start Copy New File"
+		xcopy ($SolutionDir + $content) $bundle /Y
+		xcopy ($ProjectDir + "*.*") $bundle /Y /I /E /R
+		Write-Host "************ Oh my got ! Copy Complete! Chuongmep.com ＼（＾_＾）／"
+	}
+	
 }
 else
 {
